Revert "move local notification + fullscreen message classes from compatibility to core 2.0"
This reverts commit 9b115142aa9bf998aa75efb42e46c25c173d1869.
